ref: rename user entity files name with prefix
